+++ packages/office-ui-fabric-react/src/components/Checkbox/Checkbox.tsx
@@ -1,17 +1,15 @@
 import * as React from 'react';
 import {
   BaseComponent,
-  autobind,
-  getId
+  getId,
+  createRef,
+  customizable
 } from '../../Utilities';
 import { Icon } from '../../Icon';
 import {
   ICheckbox,
   ICheckboxProps,
 } from './Checkbox.types';
-import {
-  customizable
-} from '../../Utilities';
 import {
   ICheckboxClassNames,
   getClassNames
@@ -29,7 +27,7 @@ export class Checkbox extends BaseComponent<ICheckboxProps, ICheckboxState> impl
     boxSide: 'start'
   };
 
-  private _checkBox: HTMLInputElement;
+  private _checkBox = createRef<HTMLInputElement>();
   private _id: string;
   private _classNames: ICheckboxClassNames;
 
@@ -101,7 +99,7 @@ export class Checkbox extends BaseComponent<ICheckboxProps, ICheckboxState> impl
         { ...(checked !== undefined && { checked }) }
         { ...(defaultChecked !== undefined && { defaultChecked }) }
         disabled={ disabled }
-        ref={ this._resolveRef('_checkBox') }
+        ref={ this._checkBox }
         name={ name }
         id={ this._id }
         role='checkbox'
@@ -133,13 +131,12 @@ export class Checkbox extends BaseComponent<ICheckboxProps, ICheckboxState> impl
   }
 
   public focus(): void {
-    if (this._checkBox) {
-      this._checkBox.focus();
+    if (this._checkBox.value) {
+      this._checkBox.value.focus();
     }
   }
 
-  @autobind
-  private _onFocus(ev: React.FocusEvent<HTMLInputElement>): void {
+  private _onFocus = (ev: React.FocusEvent<HTMLInputElement>): void => {
     const { inputProps } = this.props;
 
     if (inputProps && inputProps.onFocus) {
@@ -147,8 +144,7 @@ export class Checkbox extends BaseComponent<ICheckboxProps, ICheckboxState> impl
     }
   }
 
-  @autobind
-  private _onBlur(ev: React.FocusEvent<HTMLInputElement>): void {
+  private _onBlur = (ev: React.FocusEvent<HTMLInputElement>): void => {
     const { inputProps } = this.props;
 
     if (inputProps && inputProps.onBlur) {
@@ -156,8 +152,7 @@ export class Checkbox extends BaseComponent<ICheckboxProps, ICheckboxState> impl
     }
   }
 
-  @autobind
-  private _onClick(ev: React.FormEvent<HTMLButtonElement>) {
+  private _onClick = (ev: React.FormEvent<HTMLButtonElement>): void => {
     const { disabled, onChange } = this.props;
     const { isChecked } = this.state;
     ev.preventDefault();
@@ -174,8 +169,7 @@ export class Checkbox extends BaseComponent<ICheckboxProps, ICheckboxState> impl
     }
   }
 
-  @autobind
-  private _onRenderLabel(props: ICheckboxProps) {
+  private _onRenderLabel = (props: ICheckboxProps): JSX.Element | null => {
     const { label } = props;
 
     return label ? (