+++ packages/office-ui-fabric-react/src/components/pickers/Suggestions/Suggestions.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import {
   BaseComponent,
   css,
-  autobind,
   KeyCodes
 } from '../../../Utilities';
 import { CommandButton, IconButton, IButton } from '../../../Button';
@@ -193,8 +192,7 @@ export class Suggestions<T> extends BaseComponent<ISuggestionsProps<T>, ISuggest
   /**
    * Returns true if the event was handled, false otherwise
    */
-  @autobind
-  public tryHandleKeyDown(keyCode: number, currentSuggestionIndex: number): boolean {
+  public tryHandleKeyDown = (keyCode: number, currentSuggestionIndex: number): boolean => {
     let isEventHandled = false;
     let newSelectedActionType = null;
     const currentSelectedAction = this.state.selectedActionType;
@@ -304,7 +302,7 @@ export class Suggestions<T> extends BaseComponent<ISuggestionsProps<T>, ISuggest
 
   // TODO get the element to scroll into view properly regardless of direction.
   public scrollSelected() {
-    if (this._selectedElement) {
+    if (this._selectedElement && this._selectedElement.scrollIntoView !== undefined) {
       this._selectedElement.scrollIntoView(false);
     }
   }
@@ -352,41 +350,35 @@ export class Suggestions<T> extends BaseComponent<ISuggestionsProps<T>, ISuggest
       </div>);
   }
 
-  @autobind
-  private _getMoreResults() {
+  private _getMoreResults = (): void => {
     if (this.props.onGetMoreResults) {
       this.props.onGetMoreResults();
     }
   }
 
-  @autobind
-  private _forceResolve() {
+  private _forceResolve = (): void => {
     if (this.props.createGenericItem) {
       this.props.createGenericItem();
     }
   }
 
-  @autobind
-  private _shouldShowForceResolve() {
+  private _shouldShowForceResolve = (): boolean => {
     return this.props.showForceResolve ? this.props.showForceResolve() : false;
   }
 
-  @autobind
-  private _onClickTypedSuggestionsItem(item: T, index: number): (ev: React.MouseEvent<HTMLElement>) => void {
+  private _onClickTypedSuggestionsItem = (item: T, index: number): (ev: React.MouseEvent<HTMLElement>) => void => {
     return (ev: React.MouseEvent<HTMLElement>): void => {
       this.props.onSuggestionClick(ev, item, index);
     };
   }
 
-  @autobind
-  private _refocusOnSuggestions(keyCode: number): void {
+  private _refocusOnSuggestions = (keyCode: number): void => {
     if (typeof this.props.refocusSuggestions === 'function') {
       this.props.refocusSuggestions(keyCode);
     }
   }
 
-  @autobind
-  private _onRemoveTypedSuggestionsItem(item: T, index: number): (ev: React.MouseEvent<HTMLElement>) => void {
+  private _onRemoveTypedSuggestionsItem = (item: T, index: number): (ev: React.MouseEvent<HTMLElement>) => void => {
     return (ev: React.MouseEvent<HTMLElement>): void => {
       const onSuggestionRemove = this.props.onSuggestionRemove!;
       onSuggestionRemove(ev, item, index);