+++ packages/styling/CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log - @uifabric/styling
 
-This log was last generated on Fri, 23 Feb 2018 03:05:53 GMT and should not be manually modified.
+This log was last generated on Mon, 19 Mar 2018 10:27:55 GMT and should not be manually modified.
+
+## 5.20.0
+Mon, 19 Mar 2018 10:27:55 GMT
+
+### Minor changes
+
+- ThemePrimary: Updating this color along with an Office branding update.
+
+### Patches
+
+- High contrast mode for focus styling.
+
+## 5.19.1
+Mon, 12 Mar 2018 06:29:20 GMT
+
+### Patches
+
+- By default, try to use the system font before downloading a web font
 
 ## 5.19.0
 Fri, 23 Feb 2018 03:05:53 GMT